+++ GBIF/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.1.1`
+
+- **fixed** an issue where the delimiter was not properly infered by CSV [#501]
+- **added** a `path` keyword to `GBIF.download` to store downloaded files in a specific path
+
 ## `v1.1.0`
 
 - **added** support for the download API [#308]